requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings>=2.0  
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart